Practical 2: Sum of two numbers
Description:

Importing Scanner:
The program uses the Scanner class from the java.util package to allow the user to input numbers.
Class and Method Declaration:

The class is named Practical2.
The main method is the entry point of the program.
Taking Input:

A Scanner object named input is created to read user input from the console.
The program prompts the user with: "Enter the first number :". The inputted integer is stored in the variable num1.
Similarly, the user is prompted with: "Enter the second number :". The second inputted integer is stored in the variable num2.
Calculating the Sum:

The program calculates the sum of num1 and num2 and stores the result in the variable sum.